+++ source/Examples/Resources/SqlRole/5-MembersToExcludeInServerRole.ps1
@@ -18,7 +18,7 @@ Configuration Example
 
     node localhost
     {
-        SqlServerRole 'Drop_ServerRole_AdminSqlforBI'
+        SqlRole 'Drop_ServerRole_AdminSqlforBI'
         {
             Ensure               = 'Present'
             ServerRoleName       = 'AdminSqlforBI'